+++ src/main/java/com/mclegoman/perspective/mixin/client/zoom/MouseMixin.java
@@ -7,11 +7,15 @@
 
 package com.mclegoman.perspective.mixin.client.zoom;
 
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import com.mclegoman.perspective.client.config.ConfigHelper;
 import com.mclegoman.perspective.client.data.ClientData;
 import com.mclegoman.perspective.client.zoom.Zoom;
 import com.mclegoman.perspective.common.data.Data;
 import net.minecraft.client.Mouse;
+import net.minecraft.client.util.GlfwUtil;
+import net.minecraft.client.util.SmoothUtil;
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
@@ -57,5 +61,13 @@ public abstract class MouseMixin {
         } catch (Exception error) {
             Data.PERSPECTIVE_VERSION.getLogger().warn("{} An error occurred whilst trying to Mouse$onMouseButton: {}", Data.PERSPECTIVE_VERSION.getLoggerPrefix(), error);
         }
+    }
+    @ModifyExpressionValue(method = "updateMouse", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/option/Perspective;isFirstPerson()Z"))
+    private boolean perspective$isFirstPerson(boolean isFirstPerson) {
+        return ConfigHelper.getConfig("zoom_camera_mode").equals("spyglass") ? (isFirstPerson || Zoom.isZooming()) : isFirstPerson;
+    }
+    @ModifyExpressionValue(method = "updateMouse", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/network/ClientPlayerEntity;isUsingSpyglass()Z"))
+    private boolean perspective$isUsingSpyglass(boolean isUsingSpyglass) {
+        return ConfigHelper.getConfig("zoom_camera_mode").equals("spyglass") ? (isUsingSpyglass || Zoom.isZooming()) : isUsingSpyglass;
     }
 }